lab05: updated the instructions for building U-Boot in Ubuntu 22.04.2.
TO DO: Linux cannot be built without sudo, due to mknod requiring
super user permissions. Furthermore, if the /bin/busybox doesn't
have root:root ownership, the kernel boots to a normal user prompt
(denoted with $) and no modules could be modprobe-ed. If the ownership is
root:root, then the system works as expected from the root terminal
(denoted with #). This also has to be fixed.